gateway/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
pydantic==2.5.2
python-dotenv==1.0.0
openai==1.3.7
anthropic==0.7.8
boto3==1.34.10
aiohttp==3.9.1
pydantic-settings==2.1.0
tenacity==8.2.3
asyncpg==0.29.0
redis==5.0.1
psutil==5.9.7

# OpenTelemetry for distributed tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-otlp==1.21.0
